feat/#69: progress에서 strikeDay 증가할 때 Level 변경 시 DB에 저장
